give search page sub-components their own files
